requirements.txt
asgiref==3.8.1
channels==4.1.0
diff-match-patch==20241021
Django==5.0.4
django-autoslug==1.9.9
django-bootstrap5==25.1
django-ckeditor==6.7.2
django-csp==3.7
django-environ==0.12.0
django-feather==0.3.0
django-filter==25.1
django-formset==1.7.6
django-import-export==4.1.1
django-js-asset==3.1.2
django-model-utils==5.0.0
django-notifications-hq==1.8.3
django-permissions-policy==4.25.0
django-phonenumber-field==8.0.0
fontawesomefree==6.6.0
import-export==0.3.1
jsonfield==3.1.0
phonenumbers==8.13.48
pillow==11.2.1
pytz==2025.2
setuptools==78.1.0
sqlparse==0.5.3
swapper==1.4.0
tablib==3.5.0
